+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-06-15 09:55:49
-// Commit: feat(i18n): add async i18n stream provider
+// Updated at: 2026-06-15 10:43:39
+// Commit: test(i18n): implement widget integration tests for i18n
 
 class i18nComponentState {
-  final String commitRef = '1961';
-  final String lastMessage = 'feat(i18n): add async i18n stream provider';
-  final DateTime timestamp = DateTime.parse('2026-06-15T09:55:49+05:00');
+  final String commitRef = '1965';
+  final String lastMessage = 'test(i18n): implement widget integration tests for i18n';
+  final DateTime timestamp = DateTime.parse('2026-06-15T10:43:39+05:00');
 }